requirements.txt
pip==22.3.1
Django==4.2.8
django-cryptography==1.1
django-import-export==3.3.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1